APILIB-29: Extract base API classes to provide different level of integration.
