+++ infermedica_api/connectors/__init__.py
@@ -1,0 +1,2 @@
+from .common import APIConnector, SEARCH_FILTERS
+from .v2 import APIv2Connector, APIv2ModelConnector
